+++ lib/rpmts.c
@@ -68,8 +68,9 @@ extern int statvfs (const char * file, /*@out@*/ struct statvfs * buf)
 /*@access FD_t @*/		/* XXX void * arg */
 /*@access rpmdb @*/		/* XXX db->db_chrootDone, NULL */
 
-/*@access rpmps @*/
 /*@access rpmDiskSpaceInfo @*/
+/*@access rpmKeyring @*/
+/*@access rpmps @*/
 /*@access rpmsx @*/
 /*@access rpmte @*/
 /*@access rpmtsi @*/
@@ -685,7 +686,9 @@ rpmts rpmtsFree(rpmts ts)
 
     if (_rpmts_macros) {
 	const char ** av = NULL;
+/*@-globs@*/	/* Avoid rpmGlobalMcroContext et al. */
 	(void)rpmGetMacroEntries(NULL, NULL, 1, &av);
+/*@=globs@*/
 	argvPrint("macros used", av, NULL);
 	av = argvFree(av);
     }
@@ -753,6 +756,8 @@ rpmVSFlags rpmtsVSFlags(/*@unused@*/ rpmts ts)
 }
 
 rpmVSFlags rpmtsSetVSFlags(/*@unused@*/ rpmts ts, rpmVSFlags vsflags)
+	/*@globals pgpDigVSFlags @*/
+	/*@modifies pgpDigVSFlags @*/
 {
     rpmVSFlags ovsflags;
     ovsflags = pgpDigVSFlags;
@@ -1048,7 +1053,7 @@ int rpmtsInitDSI(const rpmts ts)
 	dsi->f_fsid = sfb.f_fsid;
 #endif
 	dsi->f_flag = sfb.f_flag;
-	dsi->f_favail = sfb.f_favail;
+	dsi->f_favail = (long long) sfb.f_favail;
 	dsi->f_namemax = sfb.f_namemax;
 #elif defined(__APPLE__) && defined(__MACH__) && !defined(_SYS_STATVFS_H_)
 	dsi->f_fsid = 0; /* "Not meaningful in this implementation." */
@@ -1059,17 +1064,17 @@ int rpmtsInitDSI(const rpmts ts)
 #endif
 
 	dsi->f_bsize = sfb.f_bsize;
-	dsi->f_blocks = sfb.f_blocks;
-	dsi->f_bfree = sfb.f_bfree;
-	dsi->f_files = sfb.f_files;
-	dsi->f_ffree = sfb.f_ffree;
+	dsi->f_blocks = (unsigned long long)sfb.f_blocks;
+	dsi->f_bfree = (unsigned long long)sfb.f_bfree;
+	dsi->f_files = (unsigned long long)sfb.f_files;
+	dsi->f_ffree = (unsigned long long)sfb.f_ffree;
 
 	dsi->bneeded = 0;
 	dsi->ineeded = 0;
 #ifdef STATFS_HAS_F_BAVAIL
-	dsi->f_bavail = sfb.f_bavail ? sfb.f_bavail : 1;
+	dsi->f_bavail = (long long)(sfb.f_bavail ? sfb.f_bavail : 1);
 	if (sfb.f_ffree > 0 && sfb.f_files > 0 && sfb.f_favail > 0)
-	    dsi->f_favail = sfb.f_favail;
+	    dsi->f_favail = (long long)sfb.f_favail;
 	else	/* XXX who knows what evil lurks here? */
 	    dsi->f_favail = !(sfb.f_ffree == 0 && sfb.f_files == 0)
 				?  (signed long long) sfb.f_ffree : -1;
@@ -1200,7 +1205,9 @@ void * rpmtsNotify(rpmts ts, rpmte te,
 	/*@-type@*/ /* FIX: cast? */
 	/*@-noeffectuncon @*/ /* FIX: check rc */
 	if (te) {
+/*@-mods@*/	/* XXX noisy in transaction.c */
 	    h = headerLink(te->h);
+/*@=mods@*/
 	    cbkey = rpmteKey(te);
 	} else {
 	    h = NULL;
